Create README.md
ReadMe - SimpleWebsite:
This is a simple website where we created using HTML, CSS, JavaScript and this will be interpreted by the browser such as Chrome.
HTML: is the backbone of the website where we put the main elements or content of the website.
CSS: is responsible for the style and layout of the website. Its’ role is to design the website in code efficient manner with less HTML coding.
JavaScript: is responsible for the logic to make the functions work and connect the User Interface element into different uses of the website.
With JavaScript, we used anchor tags to navigate pages.